% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{square}
\alias{square}
\title{Daily water balance of the basin}
\usage{
square(x)
}
\arguments{
\item{x}{plot is based on the timeloop.dat file present in output folder of J2K model}
}
\value{
Plot with the daily water balance
}
\description{
Plots the daily storage
}
\examples{
J2K_WatBalplot <- plot(x=years, y= daily_water_storage_in_basin)

}
